ready 3.5

+++ afanasev.yaroslav/P3/main.cpp
@@ -5,7 +5,7 @@
 
 namespace afanasev
 {
-  long long CNT_LOC_MIN(const long long *mtx, const size_t r, const size_t c)
+  long long CNT_LOC_MIN(const long long *mtx, const long long r, const long long c)
   {
     bool flag = 1;
     long long n = 0;
@@ -37,7 +37,7 @@ namespace afanasev
     return ans;
   }
 
-  long long CNT_LOC_MAX(const long long *mtx, const size_t r, const size_t c)
+  long long CNT_LOC_MAX(const long long *mtx, const long long r, const long long c)
   {
     bool flag = 1;
     long long n = 0;
@@ -81,7 +81,7 @@ int main (int argc, char ** argv)
       return 1;
     }
 
-    size_t r = 0, c = 0;
+    long long r = 0, c = 0;
     std::ifstream input(argv[2]);
     input >> r >> c;
     if (input.fail() || r < 0 || c < 0)
